refactor: Move shared/store.js -> gui/app/models/store.js
