+++ lib/gui/app/app.js
@@ -33,7 +33,7 @@ const EXIT_CODES = require('../../shared/exit-codes')
 const messages = require('../../shared/messages')
 const s3Packages = require('../../shared/s3-packages')
 const release = require('../../shared/release')
-const store = require('../../shared/store')
+const store = require('./models/store')
 const errors = require('../../shared/errors')
 const packageJSON = require('../../../package.json')
 const flashState = require('../../shared/models/flash-state')
