+++ lib/gui/app/models/settings.js
@@ -23,7 +23,7 @@
 const _ = require('lodash')
 const Bluebird = require('bluebird')
 const localSettings = require('./local-settings')
-const store = require('../../../shared/store')
+const store = require('./store')
 const errors = require('../../../shared/errors')
 
 /**
