+++ lib/gui/app/models/store.js
@@ -20,13 +20,13 @@ const Immutable = require('immutable')
 const _ = require('lodash')
 const redux = require('redux')
 const uuidV4 = require('uuid/v4')
-const constraints = require('./drive-constraints')
-const supportedFormats = require('./supported-formats')
-const errors = require('./errors')
-const release = require('./release')
-const fileExtensions = require('./file-extensions')
-const utils = require('./utils')
-const packageJSON = require('../../package.json')
+const constraints = require('../../../shared/drive-constraints')
+const supportedFormats = require('../../../shared/supported-formats')
+const errors = require('../../../shared/errors')
+const release = require('../../../shared/release')
+const fileExtensions = require('../../../shared/file-extensions')
+const utils = require('../../../shared/utils')
+const packageJSON = require('../../../../package.json')
 
 /**
  * @summary Verify and throw if any state fields are nil
