+++ lib/gui/app/pages/main/controllers/flash.js
@@ -25,7 +25,7 @@ const notification = require('../../../os/notification')
 const exceptionReporter = require('../../../modules/exception-reporter')
 const imageWriter = require('../../../modules/image-writer')
 const path = require('path')
-const store = require('../../../../../shared/store')
+const store = require('../../../models/store')
 const constraints = require('../../../../../shared/drive-constraints')
 const availableDrives = require('../../../../../shared/models/available-drives')
