+++ lib/shared/models/selection-state.js
@@ -17,7 +17,7 @@
 'use strict'
 
 const _ = require('lodash')
-const store = require('../store')
+const store = require('../../gui/app/models/store')
 const availableDrives = require('./available-drives')
 
 /**
