+++ tests/gui/models/settings.spec.js
@@ -19,7 +19,7 @@
 const m = require('mochainon')
 const _ = require('lodash')
 const Bluebird = require('bluebird')
-const store = require('../../../lib/shared/store')
+const store = require('../../../lib/gui/app/models/store')
 const settings = require('../../../lib/gui/app/models/settings')
 const localSettings = require('../../../lib/gui/app/models/local-settings')
